debug nn2 cppn decoding and bodyplantesting_nn2

+++ EAFramework/include/ARE/morphology_descriptors.hpp
@@ -18,7 +18,7 @@ struct morphology_constants{
     static constexpr int voxel_multiplier = 22;
     static constexpr float voxel_real_size = 0.0198;//voxel_size * voxel_multiplier;
     static constexpr int matrix_size = 12; //264 / voxel_multiplier;
-    static constexpr int shape_scale_value = 0.0198;//voxel_real_size; // results into 23.76x23.76x23.76 cm³ - in reality is 28x28x25 cm³
+    static constexpr float shape_scale_value = 0.0198;//voxel_real_size; // results into 23.76x23.76x23.76 cm³ - in reality is 28x28x25 cm³
     static constexpr int voxels_number = 1728;//matrix_size*matrix_size*matrix_size;
     static constexpr float matrix_size_m = 0.2376; //matrix_size * voxel_real_size;
     static constexpr short int empty_voxel=0;
@@ -51,7 +51,6 @@ struct CartDesc
     Eigen::VectorXd getCartDesc(){
         Eigen::VectorXd cartDesc(8);
         cartDesc(0) = robotWidth / morph_const::dimension_limit;
-
         cartDesc(1) = robotDepth / morph_const::dimension_limit;
         cartDesc(2) = robotHeight / morph_const::dimension_limit;
         cartDesc(3) = (double) voxelNumber / morph_const::voxels_number;
